refactoring (#55)
