+++ utils/web3client.py
@@ -6,6 +6,8 @@ import web3.types
 import requests
 import eth_account.signers.local
 
+from utils import helpers
+
 
 class NeonWeb3Client:
     def __init__(self, proxy_url: str, chain_id: int, session: tp.Optional[tp.Any] = None):
@@ -149,3 +151,30 @@ class NeonWeb3Client:
         instruction_tx = self._web3.eth.account.sign_transaction(transaction, account.key)
         signature = self._web3.eth.send_raw_transaction(instruction_tx.rawTransaction)
         return self._web3.eth.wait_for_transaction_receipt(signature)
+
+    def deploy_and_get_contract(
+        self,
+        contract_name: str,
+        version: str,
+        account: tp.Optional[eth_account.signers.local.LocalAccount] = None,
+        constructor_args: tp.Optional[tp.Any] = None,
+        gas: tp.Optional[int] = 0,
+    ) -> tp.Tuple[tp.Any, web3.types.TxReceipt]:
+        if account is None:
+            raise Exception("account parameter is None")
+        
+        contract_interface = helpers.get_contract_interface(contract_name, version)
+
+        contract_deploy_tx = self.deploy_contract(
+            account,
+            abi=contract_interface["abi"],
+            bytecode=contract_interface["bin"],
+            constructor_args=constructor_args,
+            gas=gas,
+        )
+
+        contract = self.eth.contract(
+            address=contract_deploy_tx["contractAddress"], abi=contract_interface["abi"]
+        )
+
+        return contract, contract_deploy_tx
